先月の稼働時間見れるようにする
